requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
pydantic==1.10.9
numpy==1.24.3
python-multipart